finish adding additional tests
