+++ .eslintignore
@@ -2,3 +2,4 @@ dist/
 lib/
 node_modules/
 jest.config.js
+__tests__/
