+++ jest.config.js
@@ -1,5 +1,6 @@
 module.exports = {
 	clearMocks: true,
+	resetModules: true,
 	moduleFileExtensions: ['js', 'ts'],
 	testMatch: ['**/*.test.ts'],
 	transform: {
@@ -8,5 +9,6 @@ module.exports = {
 	verbose: true,
 	collectCoverage: true,
 	collectCoverageFrom: ['src/**/*.{js,ts}'],
+	coveragePathIgnorePatterns: ['/node_modules/', '<rootDir>/src/main.ts'],
 	coverageReporters: ['clover', 'cobertura', ['text', {skipFull: true}]]
   }
